app/requirements-docs.txt
alembic
annotated-types
anyio
async-timeout
asyncpg
certifi
cfgv
click
colorama
distlib
exceptiongroup
fastapi
filelock
greenlet
h11
identify
idna
Mako
MarkupSafe
nodeenv
platformdirs
pre-commit
pydantic
pydantic-settings
pydantic_core
python-dotenv
PyYAML
sniffio
SQLAlchemy
starlette
typing_extensions
urllib3
uvicorn
virtualenv
requests
